新增Domain包裝業務資料與傳遞資料
