+++ TeaTime.Api/Domains/Store/Store.cs
@@ -1,7 +1,8 @@
-﻿namespace TeaTime.Api.Models
+﻿namespace TeaTime.Api.Domains.Store
 {
-    public class StoreDTO
+    public class Store
     {
+        // 業務處理需要的資料
         public long Id { get; set; }
 
         public string Name { get; set; } = string.Empty;
@@ -9,5 +10,6 @@
         public string PhoneNumber { get; set; } = string.Empty;
 
         public string MenuUrl { get; set; } = string.Empty;
+
     }
 }
