+++ TeaTime.Api/Domains/Store/StoreDTO.cs
@@ -1,15 +1,12 @@
-﻿namespace TeaTime.Api.Models
+﻿namespace TeaTime.Api.Domains.Store
 {
-    public class Store
+    public class StoreDTO
     {
-        public long Id { get; set; }
-
+        // 創建商店時所需的數據
         public string Name { get; set; } = string.Empty;
 
         public string PhoneNumber { get; set; } = string.Empty;
 
         public string MenuUrl { get; set; } = string.Empty;
-        public string? Secret { get; set; }
-
     }
 }
