+++ TeaTime.Api/Domains/TeaTimeContext.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using TeaTime.Api.Models;
 
 namespace TeaTime.Api.Models
 {
